Fix inconsistent results produced on different platforms
Fix a bug that resulted in different results being returned when running
stripepy call on different platform (and potentially even when using
different versions of numpy and pandas).

The bug was caused by not forcing numpy and pandas to use stable
sorting algorithms.
This resulted in slightly different results produced on macOS
and Linux (Windows and Linux produced consistent results).