Optimize overall performance on macOS and Windows
Allocating shared memory and initializing a new pool of processes
appears to be much slower on macOS and Windows than on Linux.

Given that chromosomes in .[m]cool/hic files are often sorted by size,
and that memory requirements of sparse Hi-C matrices are roughly
proportional to the size of a chromosome, on these platforms we
can try to reuse shared memory that has already been allocated.

On Linux, this trick does not seem to work, as processes are not always
notified that the main process wrote some new data to shared memory.
Thus, on Linux, we always allocate a new chunk of shared memory and
re-initialize the pool of processes for every chromosome.
This is ok as the overhead of doing so is negligible.