Reworked matrix slicing in Stripe class
Up until now, slicing the matrix resulted in a rectangular sub-matrix with
some entries referring to the incorrect half of the matrix (lower or upper).
Commit eae8489 updated the logic such that entries from the wrong half of
the matrix are discarded before computing the descriptors. This introduced
a potential problem when the stripe height is similar to the window size,
as in this case descriptors on one of the neighborhoods are computed using
a very small (if any) number of entries. This commit changes the logic used
to calculate the indices to slice the matrix such that after discarding
entries located in the wrong part of the matrix, we are left with the same
number of entries (except for stripes located at the boundary of chromosomes).
While before this commit we were padding just horizontally, we now pad also
along the vertical direction.

Co-authored-by: Andrea Raffo <52046472+rea1991@users.noreply.github.com>
Co-authored-by: Roberto Rossini <71787608+robomics@users.noreply.github.com>