Reworked matrix slicing in Stripe class one more time
The logic introduced with edd0bd7 did not work as expected when we
failed to expand the stripe matrix in the proximity of the boundaries.
This commit addresses this shortcoming.
We now use nan to mark masked/invalid entries, and then we compute
the biodescriptors using, e.g., np.nanmean, np.nanstd and so on.

Co-authored-by: Andrea Raffo <52046472+rea1991@users.noreply.github.com>
Co-authored-by: Roberto Rossini <71787608+robomics@users.noreply.github.com>