Introduce v3 of the schema of files generated by stripepy call
The main difference between v2 and v3 concernis how biodescriptors are stored.
In v2 we used to store the mean number of interactions of the left and
right neighborhood.
However, the changes introduced by 56ac76f and 14001e4 changed how the
outer biodescriptors are computed in such a way that computing the outer
mean as the average of the left and right neighborhoods is no longer correct.
This, in v3 we store the sum and size of the left and right neighborhoods.

Another minor difference is that /index/ut/chrom_offsets_pd is now a
symlink to /index/lt/chrom_offsets_pd (given that the two indices are
always supposed to be identical).